requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.6.1
Django==2.2.1
django-rest-knox==4.0.1
djangorestframework==3.9.3
idna==2.8
isort==4.3.18
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.24.3
wrapt==1.11.1
